requirements.txt
cffi==1.14.0
configparser==4.0.2
cryptography==2.9.2
pycparser==2.20
six==1.14.0
